add type to status icon
